Delete posts module and move collection, filters and svaed searches to map module
