+++ legacy/app/map/collections/editor.directive.js
