+++ legacy/app/map/collections/editor.html
