+++ legacy/app/map/collections/listing.directive.js
