+++ legacy/app/map/collections/mode-context.directive.js
