+++ legacy/app/map/post-toolbar/filters/active-search-filters.directive.js
