+++ legacy/app/map/post-toolbar/filters/filter-category.directive.js
