+++ legacy/app/map/post-toolbar/filters/filter-location.directive.js
