+++ legacy/app/map/post-toolbar/filters/filter-post-sorting-options.directive.js
