+++ legacy/app/map/post-toolbar/filters/filter-transformers.service.js
