+++ legacy/app/map/post-toolbar/filters/filter-visible-to.directive.js
