+++ legacy/app/map/savedsearches/editor-directive.js
