+++ legacy/app/map/savedsearches/mode-context.directive.js
