+++ legacy/gulpfile.babel.js
@@ -47,7 +47,6 @@ let paths = {
     entry: {
         "ushahidi-legacy-app": ['babel-polyfill', path.join(__dirname, root, 'bootstrap.js')],
         "data": path.join(__dirname, root, '/data/data-module.js'),
-        "posts": path.join(__dirname, root, '/posts/posts-module.js'),
         "activity": path.join(__dirname, root, '/activity/activity-module.js'),
         "settings": path.join(__dirname, root, '/settings/settings.module.js')
     },
